submodules/webgenagent/apps/agent/requirements.txt
fastapi==0.115.0
uvicorn>=0.31.1
httpx>=0.27.1,<0.28
boto3==1.34.153
python-dotenv>=1.1.1,<2.0.0
orjson==3.10.7
sse-starlette==1.8.2
supabase==2.5.1
langgraph>=0.0.10
langchain-google-genai
langchain-anthropic
langchain-openai
pydantic>=2.0.0
python-multipart
